fix customer and staff

+++ DATNBackEnd/src/main/java/com/example/demo/entity/Color.java
@@ -25,6 +25,6 @@ public class Color {
     private String name;
     @Basic
     @Column(name = "status")
-    private byte status;
+    private int status;
 }
